Preliminary annotation tool for StaBi content
StaBi content can be found here:
https://digital.staatsbibliothek-berlin.de/suche?category%5B0%5D=Tibetica

This Jupyter Notebook allows to annoate images found there and saving
the corresponding annotations in a JSON file annotations.json.
If this JSON file is kept in this git repository and updated, we'll be
able to utilise all our annotations.

The notebook reads the annotations.json in the beginning to be able to
work on top of existing work. This means it's also possible to edit
annotations of other people - for example when the notebook has a nifty
new feature that previously wasn't existing.

More documentation on the notebook likely in the Discord channel.